feat. gameRoomPage 리펙토링 #128 fix. 실행할때 TypeError 해결 (코드를 다시 롤백하여 해결) #136
